refactor: 공통되는 부분 components폴더로 분리
